# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Notify Module",
        "contact": {
            "email": "pvalente@onesource.pt"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/emailGroup/{email_group_key}": {
            "get": {
                "tags": [
                    "Email Group"
                ],
                "summary": "Show a Email Group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email_group_key",
                        "in": "path",
                        "description": "Email Group Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show the Email Group data",
                        "schema": {
                            "$ref": "#/definitions/emailGroupReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Email Group not Found",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve Email Group",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Email Group"
                ],
                "summary": "Update an Email Group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Body",
                        "in": "body",
                        "description": "Email Group Data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emailGroupCreate"
                        }
                    },
                    {
                        "name": "email_group_key",
                        "in": "path",
                        "description": "Email Group Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated Email Group",
                        "schema": {
                            "$ref": "#/definitions/emailGroupReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Email Group not Found",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to update Email Group",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Email Group"
                ],
                "summary": "Delete an Email Group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email_group_key",
                        "in": "path",
                        "description": "Email Group Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emailGroupDeleteReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Email Group not Found",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to delete Email Group",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    }
                }
            }
        },
        "/emailGroup": {
            "post": {
                "tags": [
                    "Email Group"
                ],
                "summary": "Create a Email Group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "EmailGroup",
                        "in": "body",
                        "description": "Email Group Data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emailGroupCreate"
                        }
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "the newly created Email Group",
                        "schema": {
                            "$ref": "#/definitions/emailGroupReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Email Group not found",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to store Email Group",
                        "schema": {
                            "$ref": "#/definitions/emailGroupErrorDefault"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "emailGroupErrorDefault": {
            "properties": {
                "error": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "emailGroupCreate": {
            "type": "object",
            "allOf": [
                {
                    "required": [
                        "site_key"
                    ],
                    "properties": {
                        "site_key": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                }
            ]
        },
        "emailGroupReply": {
            "type": "object",
            "allOf": [
                {
                    "properties": {
                        "id": {
                            "type": "integer",
                            "format": "integer"
                        },
                        "email_group_key": {
                            "type": "string",
                            "format": "string"
                        },
                        "site_key": {
                            "type": "string",
                            "format": "string"
                        },
                        "created_at": {
                            "type": "string",
                            "format": "date"
                        },
                        "updated_at": {
                            "type": "string",
                            "format": "date"
                        },
                        "deleted_at": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                }
            ]
        },
        "emailGroupDeleteReply": {
            "properties": {
                "string": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
...
    },
...
}